+++ src/front-end/pages/Checkout.jsx
@@ -1,0 +1,9 @@
+const Checkout = () => {
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default Checkout
