+++ src/front-end/pages/OrderHistory.jsx
@@ -1,0 +1,9 @@
+const OrderHistory = () => {
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default OrderHistory
