+++ src/front-end/pages/Products.jsx
@@ -1,0 +1,9 @@
+const Products = () => {
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default Products
