--- src/front-end/src/pages/Checkout.jsx
@@ -1,9 +1,0 @@
-const Checkout = () => {
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default Checkout
