--- src/front-end/src/pages/OrderHistory.jsx
@@ -1,9 +1,0 @@
-const OrderHistory = () => {
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default OrderHistory
