--- src/front-end/src/pages/Products.jsx
@@ -1,9 +1,0 @@
-const Products = () => {
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default Products
